requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
aria2p==0.12.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
loguru==0.7.2
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
Pyrogram==2.0.106
PySocks==1.7.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
speedtest-cli==2.1.3
TgCrypto==1.2.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
